osbackend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
kombu==5.3.1
loguru==0.7.0
numpy==1.24.4
packaging==23.1
pandas==2.0.3
pbkdf2==1.3
prompt-toolkit==3.0.39
py5paisa==0.7.2
pycryptodome==3.18.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.2
